refactor(test): drop the duplicated reassembly mirror from the merge fuzz
The reassembly-contract fuzz had hand-copied the web resolver's reassembly
logic (a mirror that could silently drift) plus the prng/mutate helpers from the
adversarial suite. Rewrite it to assert the underlying ENGINE invariant directly
— merge3's hunks rejoin to `merged`, and resolving every region to one side
loses none of that side's text — and share one copy of the fuzz scaffolding
(merge3-fuzz-helpers) across both suites. The resolver's own reassembly stays
unit-tested in apps/web; nothing here duplicates production code now.
